Feat: Adds election data import functionality
Implements functionality to import election data from uploaded files.
This includes voter lists, ballot styles, and ballot style links.
The system now uploads the file, saves it to the appropriate directory
and then executes a stored procedure to import the data into the database.
Includes the addition of connection string to the database.